moving mocks to model package
