+++ data/src/test/java/br/com/pebmed/data/pullrequest/PullRequestRepositoryImplTest.kt
@@ -1,15 +1,9 @@
 package br.com.pebmed.data.pullrequest
 
-import br.com.pebmed.data.pullRequest.PullRequestRemoteDataSource
 import br.com.pebmed.data.pullRequest.PullRequestRepositoryImpl
-import br.com.pebmed.data.pullRequest.model.PullRequestResponseModel
-import br.com.pebmed.data.pullrequest.model.mock.MockPullRequestResponseModel
-import br.com.pebmed.domain.base.CompleteResultWrapper
 import io.mockk.*
-import io.mockk.impl.annotations.MockK
 import junit.framework.Assert.assertNotNull
 import kotlinx.coroutines.runBlocking
-import org.junit.After
 import org.junit.Before
 import org.junit.Test
